refactor: reorganize tests to achieve 1:1 correspondence with application structure
- Create dedicated test files for utility modules (test_amount_utils.py, test_date_utils.py)
- Move parser tests from parsers/ subdirectory to proper application/ level
- Consolidate all model tests into test_models.py (removed test_models_integrated.py)
- Ensure every application module has corresponding test file
- Maintain full test coverage: 78 tests passing, 10.00/10 pylint score
- Clean up test structure to mirror application architecture exactly

Perfect 1:1 mapping achieved:
application/module.py → tests/unit/application/test_module.py
